// src/generated/models/Cr22e_contactosproveedorsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr22e_contactosproveedorscr22e_tipocontacto = {
  100000001: 'Ventas',
  100000002: 'Soporte',
  100000003: 'Tecnico',
  100000004: 'Administrativo',
  100000005: 'Emergencia'
} as const;
export type Cr22e_contactosproveedorscr22e_tipocontacto = keyof typeof Cr22e_contactosproveedorscr22e_tipocontacto;
export const Cr22e_contactosproveedorsstatecode = {
  0: 'Activo',
  1: 'Inactivo'
} as const;
export type Cr22e_contactosproveedorsstatecode = keyof typeof Cr22e_contactosproveedorsstatecode;
export const Cr22e_contactosproveedorsstatuscode = {
  1: 'Activo',
  2: 'Inactivo'
} as const;
export type Cr22e_contactosproveedorsstatuscode = keyof typeof Cr22e_contactosproveedorsstatuscode;

export interface Cr22e_contactosproveedorsBase {
  cr22e_activo?: boolean;
  cr22e_contactosproveedorid: string;
  cr22e_correo?: string;
  cr22e_name: string;
  cr22e_observaciones?: string;
  "cr22e_Proveedor@odata.bind"?: string;
  cr22e_puesto?: string;
  cr22e_telefono?: string;
  cr22e_tipocontacto?: Cr22e_contactosproveedorscr22e_tipocontacto;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr22e_contactosproveedorsstatecode;
  statuscode?: Cr22e_contactosproveedorsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr22e_contactosproveedors extends Cr22e_contactosproveedorsBase {
  cr22e_activoname?: string;
  cr22e_proveedorname?: string;
  cr22e_tipocontactoname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cr22e_proveedor?: object;
  _cr22e_proveedor_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}