// src/generated/models/Cr22e_serviciosproveedorsucursalsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr22e_serviciosproveedorsucursalscr22e_estadoservicio = {
  100000001: 'Activo',
  100000002: 'Inactivo',
  100000003: 'Suspendido'
} as const;
export type Cr22e_serviciosproveedorsucursalscr22e_estadoservicio = keyof typeof Cr22e_serviciosproveedorsucursalscr22e_estadoservicio;
export const Cr22e_serviciosproveedorsucursalscr22e_tiposervicio = {
  100000000: 'Internet',
  100000001: 'CCTV',
  100000002: 'Computadoras'
} as const;
export type Cr22e_serviciosproveedorsucursalscr22e_tiposervicio = keyof typeof Cr22e_serviciosproveedorsucursalscr22e_tiposervicio;
export const Cr22e_serviciosproveedorsucursalsstatecode = {
  0: 'Activo',
  1: 'Inactivo'
} as const;
export type Cr22e_serviciosproveedorsucursalsstatecode = keyof typeof Cr22e_serviciosproveedorsucursalsstatecode;
export const Cr22e_serviciosproveedorsucursalsstatuscode = {
  1: 'Activo',
  2: 'Inactivo'
} as const;
export type Cr22e_serviciosproveedorsucursalsstatuscode = keyof typeof Cr22e_serviciosproveedorsucursalsstatuscode;

export interface Cr22e_serviciosproveedorsucursalsBase {
  cr22e_activo?: boolean;
  cr22e_correosoporte?: string;
  cr22e_estadoservicio?: Cr22e_serviciosproveedorsucursalscr22e_estadoservicio;
  cr22e_fechaultimaactualizacion?: string;
  cr22e_horarioatencion?: string;
  cr22e_name: string;
  cr22e_observaciones?: string;
  "cr22e_Proveedor@odata.bind"?: string;
  cr22e_serviciosproveedorsucursalid: string;
  "cr22e_Sucursal@odata.bind"?: string;
  cr22e_telefonosoporte?: string;
  cr22e_tiposervicio?: Cr22e_serviciosproveedorsucursalscr22e_tiposervicio;
  "cr22e_ubicaciondentrosucursal@odata.bind"?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr22e_serviciosproveedorsucursalsstatecode;
  statuscode?: Cr22e_serviciosproveedorsucursalsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr22e_serviciosproveedorsucursals extends Cr22e_serviciosproveedorsucursalsBase {
  cr22e_activoname?: string;
  cr22e_estadoservicioname?: string;
  cr22e_proveedorname?: string;
  cr22e_sucursalname?: string;
  cr22e_tiposervicioname?: string;
  cr22e_ubicaciondentrosucursalname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cr22e_proveedor?: object;
  _cr22e_proveedor_value?: string;
  cr22e_sucursal?: object;
  _cr22e_sucursal_value?: string;
  cr22e_ubicaciondentrosucursal?: object;
  _cr22e_ubicaciondentrosucursal_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}